requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
sentence-transformers
pypdf
rank-bm25